pyproject.toml
[project]
name = "auction"
version = "0.1.0"
description = "Commercial Vehicle Leaseback Pricing Optimizer"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "python-multipart",
    "supabase",
    "httpx",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "structlog",
    "cachetools",
    "numpy",
    "pandas",
    "openpyxl",
    "reportlab",
    "playwright",
    "beautifulsoup4",
    "lxml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests covering full user workflows",
    "monkey: monkey/fuzz tests with random and adversarial inputs",
]

[tool.mypy]
python_version = "3.11"
strict = true
